First commit, working so far
The game seems to be working, although there are some changes that must be
made. There is no procedure for the "special finishings", i.e., ending on a
double piece, ending in a piece playable in both sides, or both.

The greatest problem that has to be addressed, though, is that many times there is more than one end to insert the piece, and this choice is simply ignored by
the game.

For the next commit, I will work on this and the next level, namely, 6-point
tournaments.