Table holds/shows what players declared NOT having
This was not implemented before, but it makes sense that the table
(which represents the "dealer"/referee of the game) should keep
track of who has declared not having pieces to play in each turn.

Having this ready will make the next step easier: calculating the
probabilities of the adversaries having each of the numbers.